Remove minification of JS and CSS
It's technically not necessary and takes too much time.

Minified files can cause problems when debugging, because they are
preferred to their non-minified siblings – so if you change a file and
don't see the result in the browser, check if a minified version of the
edited file is present and loaded instead.